pyproject.toml
[tool.poetry]
name = "nu_abda"
version = "0.1.0"
description = ""
authors = ["lgblkb <dbakhtiyarov@nu.edu.kz>"]

[tool.poetry.dependencies]
python = "^3.6.9"
wheel = "^0.34.2"
setuptools = "^46.1.3"
lgblkb_tools = "^1.1.15"
torch = "^1.5.0"
torchvision = "^0.6.0"
imgaug = "^0.4.0"
sklearn = "^0.0"
wandb = "^0.8.33"
python-box = "<4"
taskipy = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"

[tool.taskipy.tasks]
export="poetry export -f requirements.txt -o requirements.txt --without-hashes"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

